+++ R/init.R
@@ -83,16 +83,6 @@ load_gdal <- function() {
 	}
 	CPL_gdal_init()
 	register_all_s3_methods() # dynamically registers non-imported pkgs (tidyverse)
-	if (inherits(try(units::as_units("link"), silent = TRUE), "try-error"))
-		units::install_conversion_constant("m", "link", 0.201168)
-	if (inherits(try(units::as_units("us_in"), silent = TRUE), "try-error"))
-		units::install_conversion_constant("m", "us_in", 1./39.37)
-	if (inherits(try(units::as_units("ind_yd"), silent = TRUE), "try-error"))
-		units::install_conversion_constant("m", "ind_yd", 0.91439523)
-	if (inherits(try(units::as_units("ind_ft"), silent = TRUE), "try-error"))
-		units::install_conversion_constant("m", "ind_ft", 0.30479841)
-	if (inherits(try(units::as_units("ind_ch"), silent = TRUE), "try-error"))
-		units::install_conversion_constant("m", "ind_ch", 20.11669506)
 }
 
 unload_gdal <- function() {
@@ -105,9 +95,24 @@ unload_gdal <- function() {
 		Sys.setenv("GDAL_DATA"=get(".sf.GDAL_DATA", envir=.sf_cache))
 		# nocov end
 	}
-	units::remove_symbolic_unit("link")
-	units::remove_symbolic_unit("us_in")
-	units::remove_symbolic_unit("ind_yd")
-	units::remove_symbolic_unit("ind_ft")
-	units::remove_symbolic_unit("ind_ch")
+	#units::remove_symbolic_unit("link")
+	#units::remove_symbolic_unit("us_in")
+	#units::remove_symbolic_unit("ind_yd")
+	#units::remove_symbolic_unit("ind_ft")
+	#units::remove_symbolic_unit("ind_ch")
+}
+ 
+
+#' @export
+#' @name sf_project
+#' @details \code{sf_add_proj_units} loads the PROJ units `link`, `us_in`, `ind_yd`, `ind_ft`, and `ind_ch` into the udunits database, and returns \code{TRUE} invisibly on success.
+sf_add_proj_units = function() {
+	#nocov start
+	units::install_conversion_constant("m", "link", 0.201168)
+	units::install_conversion_constant("m", "us_in", 1./39.37)
+	units::install_conversion_constant("m", "ind_yd", 0.91439523)
+	units::install_conversion_constant("m", "ind_ft", 0.30479841)
+	units::install_conversion_constant("m", "ind_ch", 20.11669506)
+	invisible(TRUE)
+	#nocov end
 }